requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pytrends==4.9.2
beautifulsoup4==4.12.3

